refactor(queue): change queue model to be in models and remove unneeded queue.js
